+++ Runtime/TextureAtlas/TextureFineTuning/ColorSpace.cs
@@ -8,22 +8,28 @@ namespace net.rs64.TexTransTool.TextureAtlas.FineTuning
     [Serializable]
     public class ColorSpace : ITextureFineTuning
     {
-        public PropertyName PropertyNames;
-        public PropertySelect Select;
-        public bool Linear;
+        [Obsolete("V4SaveData", true)] public PropertyName PropertyNames = PropertyName.DefaultValue;
+        public List<PropertyName> PropertyNameList = new() { PropertyName.DefaultValue };
+        public PropertySelect Select = PropertySelect.Equal;
+        public bool Linear = true;
 
         public ColorSpace() { }
+        [Obsolete("V4SaveData", true)]
         public ColorSpace(PropertyName propertyNames, PropertySelect select, bool linear)
         {
             PropertyNames = propertyNames;
             Select = select;
             Linear = linear;
         }
-        public static ColorSpace Default => new(PropertyName.DefaultValue, PropertySelect.Equal, true);
-
+        public ColorSpace(List<PropertyName> propertyNames, PropertySelect select, bool linear)
+        {
+            PropertyNameList = propertyNames;
+            Select = select;
+            Linear = linear;
+        }
         public void AddSetting(Dictionary<string, TexFineTuningHolder> texFineTuningTargets)
         {
-            foreach (var target in FineTuningUtil.FilteredTarget(PropertyNames, Select, texFineTuningTargets))
+            foreach (var target in FineTuningUtil.FilteredTarget(PropertyNameList, Select, texFineTuningTargets))
             {
                 var tuningDataHolder = target.Value;
                 tuningDataHolder.Get<ColorSpaceData>().Linear = Linear;