+++ Runtime/TextureAtlas/TextureFineTuning/ITextureFineTuning.cs
@@ -24,9 +24,9 @@ namespace net.rs64.TexTransTool.TextureAtlas.FineTuning
 
     internal static class FineTuningUtil
     {
-        public static IEnumerable<KeyValuePair<string,TexFineTuningHolder>> FilteredTarget(string propertyNames, PropertySelect select, Dictionary<string,TexFineTuningHolder> targets)
+        public static IEnumerable<KeyValuePair<string,TexFineTuningHolder>> FilteredTarget(List<PropertyName> propertyNames, PropertySelect select, Dictionary<string,TexFineTuningHolder> targets)
         {
-            var propertyNameList = propertyNames.Split(' ');
+            var propertyNameList = propertyNames.Select(i => i.ToString()).ToHashSet();
             switch (select)
             {
                 default: