+++ Runtime/TextureAtlas/TextureFineTuning/Remove.cs
@@ -7,21 +7,26 @@ namespace net.rs64.TexTransTool.TextureAtlas.FineTuning
     [Serializable]
     public class Remove : ITextureFineTuning
     {
-        public PropertyName PropertyNames;
-        public PropertySelect Select;
+        [Obsolete("V4SaveData", true)] public PropertyName PropertyNames = PropertyName.DefaultValue;
+        public List<PropertyName> PropertyNameList = new() { PropertyName.DefaultValue };
+        public PropertySelect Select = PropertySelect.NotEqual;
 
         public Remove() { }
+        [Obsolete("V4SaveData", true)]
         public Remove(PropertyName propertyNames, PropertySelect select)
         {
             PropertyNames = propertyNames;
             Select = select;
         }
-
-        public static Remove Default => new(PropertyName.DefaultValue, PropertySelect.NotEqual);
+        public Remove(List<PropertyName> propertyNames, PropertySelect select)
+        {
+            PropertyNameList = propertyNames;
+            Select = select;
+        }
 
         public void AddSetting(Dictionary<string, TexFineTuningHolder> texFineTuningTargets)
         {
-            foreach (var target in FineTuningUtil.FilteredTarget(PropertyNames, Select, texFineTuningTargets))
+            foreach (var target in FineTuningUtil.FilteredTarget(PropertyNameList, Select, texFineTuningTargets))
             {
                 target.Value.Get<RemoveData>();
             }