+++ Runtime/TextureAtlas/TextureFineTuning/Resize.cs
@@ -9,30 +9,34 @@ namespace net.rs64.TexTransTool.TextureAtlas.FineTuning
     [Serializable]
     public class Resize : ITextureFineTuning
     {
-        [PowerOfTwo] public int Size;
-        public PropertyName PropertyNames;
-        public PropertySelect Select;
+        [PowerOfTwo] public int Size = 512;
+        [Obsolete("V4SaveData",true)] public PropertyName PropertyNames = PropertyName.DefaultValue;
+        public List<PropertyName> PropertyNameList = new() { PropertyName.DefaultValue };
+        public PropertySelect Select = PropertySelect.NotEqual;
 
         public Resize() { }
+        [Obsolete("V4SaveData",true)]
         public Resize(int size, PropertyName propertyNames, PropertySelect select)
         {
             Size = size;
             PropertyNames = propertyNames;
+            Select = select;
+        }
+        public Resize(int size, List<PropertyName> propertyNames, PropertySelect select)
+        {
+            Size = size;
+            PropertyNameList = propertyNames;
             Select = select;
 
         }
 
         public void AddSetting(Dictionary<string, TexFineTuningHolder> texFineTuningTargets)
         {
-            foreach (var target in FineTuningUtil.FilteredTarget(PropertyNames, Select, texFineTuningTargets))
+            foreach (var target in FineTuningUtil.FilteredTarget(PropertyNameList, Select, texFineTuningTargets))
             {
                 target.Value.Get<SizeData>().TextureSize = Size;
             }
         }
-
-
-        public static Resize Default => new(512, PropertyName.DefaultValue, PropertySelect.NotEqual);
-        public ITextureFineTuning GetDefault => Default;
     }
 
     internal class SizeData : ITuningData